fix: prevent firing default state queries on page load
